Add db transaction and upload feedback
Uploads are applied within a transaction in Malware.import. If
an error occurs the database is safely rolled back and no changes are
applied.

When upload errors occur .import returns an error message that is
displayed to the user. If the problem was caused by a csv row, the
line number and the entire row are displayed to help the user track
down the source of the problem